[not verified] re-organize transforms code
